requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.4        
Django==3.0.5
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
docopt==0.6.2
idna==2.10
pipreqs==0.4.10
pycparser==2.20
pymongo==3.11.3
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
urllib3==1.26.3
yarg==0.1.9
